+++ CHANGELOG.md
@@ -1,12 +1,16 @@
 # Changelog
 
-## 2.2.4 - 
+## 2.2.5 - 
 
 - Allow boolean arguments to override baked arguments [#770](https://github.com/amoffat/sh/pull/770)
 - Added .pyi typing stub
 - Improvements to str representation of `Command` [#756](https://github.com/amoffat/sh/pull/756)
 - Bugfix where rare `ProcessLookupError` can be thrown [#769](https://github.com/amoffat/sh/pull/769)
 - Various spelling corrections [#774](https://github.com/amoffat/sh/pull/774)
+
+## 2.2.4 - 6/06/25
+
+- Bugfix regression where correct `gid` was not set on the launched process
 
 ## 2.2.3 - 6/05/25